Narrow the Fortran ABI to 32 bits, and fix the compiler bug that exposed
Task #12, and it was never a decision.  SZLONG is pinned TWICE: typesize[TYREAL]
is SZLONG against libF77's r_nint taking `float *', AND lengtype() at
proc.c:951 hardcodes the length constants -- INTEGER*4 falls to `goto ret' and
returns TYLONG, INTEGER*2 gives TYSHORT, REAL*4 gives TYREAL.  So
typesize[TYLONG] must be 4 whatever the float layout said, and no other value
of SZLONG satisfies both.

The runtime's `long' therefore had to narrow, and V8's own compiler agrees it
always meant 32 bits: `# define NOLONG' at ccom/vax/macdefs.h:20.  39 libF77
files (56 `long int' sites, ef1asc_'s word-size M, outstr_'s register long, two
%ld) and 4 libI77 sites (fio.h's two typedefs, fmt.h's uint.il, nio.c's plong
and one sizeof(long)).  Byte-identical files 153 -> 112 of 154, stated rather
than hidden; PROVENANCE keeps the diff reconstructible.

The half needing judgement was what NOT to change: `long' as a host file offset
in 13 libI77 files, the `long x' locals in rd_I/wrt_I that widen a value for
icvt, and ltostr.c.  A blanket -Dlong=int breaks every one of them.

AND IT FOUND A LIVE COMPILER BUG WITHIN THE HOUR -- the sixth arm64_trunc()
site, and the only one that is a CONVERSION rather than an operator.
`fcvtzs Wd, Dn' writes a w register, which zeroes bits 63:32, and this back end
keeps a signed int sign-extended: every arithmetic site restores that with
sxtw and the conversion had none, so (int)(-2.5) was 0x00000000fffffffd.  The
tell is this port's signature -- printf("%d") reads the low half and is RIGHT
while `== -3' compares all 64 bits and is WRONG, so the failure printed
`want -3 got -3'.  Reproduced in ten lines; gencode.c calls arm64_widen() there
now.

It survived five earlier sites because arm64_trunc() switches on an OPERATOR
and a CONV is not one, so the sweep that added the unary - and ~ could not
reach it; and nothing observed it because libF77's intrinsics returned `long'
until this commit.  A width change is a way of asking a compiler questions it
has never been asked.

Three of the four new v8ccom cases were vacuous on their first draft, using an
automatic where the store re-narrows the value -- only `register' exposes it,
which is why the four operator cases beside them all say register.  The
non-firing mutation is what said so; two fire now and two are stated controls.

selfhost still 34/34, so the compiler reproduces itself byte-identically after
the codegen change.  2592 -> 2597 tests across 17 suites, 0 failed.

+++ v8/src/libI77/fmt.h
@@ -51,7 +51,7 @@ typedef union
 typedef union
 {	short is;
 	char ic;
-	long il;
+	int il;		/* PORT: the FORTRAN INTEGER arm; see fio.h */
 } uint;
 #define GET(x) if((x=(*getn)())<0) return(x)
 #define VAL(x) (x!='\n'?x:' ')
